Teste adicional e revisão.

+++ app/Http/Controllers/Api/SeriesController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\SeriesFormRequest;
 use App\Models\Series;
 use App\Repositories\SeriesRepository;
+use Illuminate\Contracts\Auth\Authenticatable;
 use Illuminate\Http\Request;
 
 class SeriesController extends Controller
@@ -65,8 +66,12 @@ class SeriesController extends Controller
         return $series;
     }
 
-    public function destroy(int $series){
-        Series::destroy($series);
+    public function destroy(int $series, Authenticatable $user){
+        //dd($user->tokenCan('series:delete'));
+        if($user->tokenCan('series:delete')){
+            Series::destroy($series);
+        }
+        
         return response()->noContent();
     }
 }
